% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getformula.R
\docType{methods}
\name{getformula}
\alias{getformula}
\alias{getformula-methods}
\alias{getformula,ergm-method}
\alias{getformula,btergm-method}
\alias{getformula,mtergm-method}
\alias{getformula,tbergm-method}
\title{Extract the formula from a model}
\usage{
getformula(x)

\S4method{getformula}{ergm}(x)

\S4method{getformula}{btergm}(x)

\S4method{getformula}{mtergm}(x)

\S4method{getformula}{tbergm}(x)
}
\arguments{
\item{x}{A fitted model.}
}
\description{
Extract the model formula from a fitted object.
}
\details{
The \code{getformula} function will extract the formula from a fitted model.
}
\section{Methods (by class)}{
\itemize{
\item \code{getformula(ergm)}: Extract the formula from an \code{ergm} object.

\item \code{getformula(btergm)}: Extract the formula from a \code{btergm} object.

\item \code{getformula(mtergm)}: Extract the formula from an \code{mtergm} object.

\item \code{getformula(tbergm)}: Extract the formula from a \code{tbergm} object.

}}
